aura-backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3

# ML and Data Science
xgboost==2.1.2
shap==0.46.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3

# Validation and serialization
pydantic==2.9.2
pydantic-settings==2.6.0

# Caching (optional for now)
# redis==5.0.8

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
